# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bn-mobile_server_internal_delivery_data_response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: API backend Belajar Ngaji (Go + Gin).
  title: BN Mobile API
  version: "1.0"
paths:
  /audio-proxy:
    get:
      description: Proxy local/allowed audio URL for client playback.
      parameters:
      - description: Audio source URL or local path
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
      summary: Audio proxy
      tags:
      - Public
...
  /prayer-times:
    get:
      description: Get prayer times by coordinates and optional date/method.
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lng
        required: true
        type: number
      - description: Date in YYYY-MM-DD (default today)
        in: query
        name: date
        type: string
      - description: Calculation method (default 2)
        in: query
        name: method
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
      summary: Get prayer times
      tags:
      - Public
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer access token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"